requirements.txt
transformers
python-dotenv
requests
torch
soundfile
datasets
nltk